Some re-structuring
