+++ web/package.json
@@ -20,6 +20,7 @@
     "angular-in-memory-web-api": "~0.3.0",
     "core-js": "^2.4.1",
     "jquery": "^3.2.1",
+    "lodash": "^4.17.4",
     "moment": "^2.18.1",
     "monet": "^0.8.10",
     "rxjs": "5.0.1",
