+++ web/src/app-container/authentication/components/authenticationComponent.ts
@@ -1,0 +1,8 @@
+import {Component} from "@angular/core";
+
+@Component({
+    selector: "ma-authentication",
+    templateUrl: "assets/target/app-container/authentication/components/authentication.ng.html",
+    styleUrls: ["assets/target/app-container/authentication/components/authentication.css"]
+})
+export class AuthenticationComponent {}
