--- web/src/app-container/components/appContainerComponent.ts
@@ -1,8 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-    selector: "management-app",
-    templateUrl: "assets/target/app-container/components/appContainer.html",
-    styleUrls: ["assets/target/app-container/components/appContainer.css"]
-})
-export class AppContainerComponent {}
