--- web/src/app-container/components/authenticationComponent.ts
@@ -1,8 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-    selector: "ma-authentication",
-    templateUrl: "assets/target/app-container/components/authentication.ng.html",
-    styleUrls: ["assets/target/app-container/components/authentication.css"]
-})
-export class AuthenticationComponent {}
