--- web/src/app-container/components/homeComponent.ts
@@ -1,8 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-    selector: "empty",
-    templateUrl: "assets/target/app-container/components/home.ng.html",
-    styleUrls: ["assets/target/app-container/components/home.css"]
-})
-export class HomeComponent {}
