+++ web/src/app-container/containers/components/appContainerComponent.ts
@@ -1,0 +1,8 @@
+import {Component} from "@angular/core";
+
+@Component({
+    selector: "management-app",
+    templateUrl: "assets/target/app-container/containers/components/appContainer.html",
+    styleUrls: ["assets/target/app-container/containers/components/appContainer.css"]
+})
+export class AppContainerComponent {}
