+++ web/src/app-container/containers/components/homeComponent.ts
@@ -1,0 +1,8 @@
+import {Component} from "@angular/core";
+
+@Component({
+    selector: "empty",
+    templateUrl: "assets/target/app-container/containers/components/home.ng.html",
+    styleUrls: ["assets/target/app-container/containers/components/home.css"]
+})
+export class HomeComponent {}
